Before adding the PIManagementDelegate (class has been created but it's empty). I'm commiting just to have the current versions of what I'm doing.
git-svn-id: svn://localhost/tmp/repo1/trunk@57 70b6eb62-b118-4aea-92db-fdb7de3615d6

+++ GOD_Console_Prototype_1/src/ddproto1/localagent/client/GlobalAgentProxyImpl.java
@@ -77,7 +77,7 @@ class GlobalAgentProxyImpl implements IGlobalAgent{
     /* (non-Javadoc)
      * @see ddproto1.localagent.client.IGlobalAgent#syncNotify(ddproto1.localagent.client.Event)
      */
-    public int syncNotify(Event e) 
+    public byte syncNotify(Event e) 
     	throws CommException
     {
         /* Acquires a connection */
